lambda-process-image/requirements.txt
Pillow == 10.4